+++ src/app/dashboard/manage/report/page.tsx
@@ -1,4 +1,4 @@
-import ReportOverview from "@/components/dashboard/report/organisms/ReportOverview";
+import ReportOverview from "@/components/dashboard/report/molecules/ReportOverview";
 import ReportList from "@/components/dashboard/report/organisms/ReportList";
 
 export default function ClubReportPage() {
